factor a few methods out into helpers.py
